Add `podman generate systemd --pull` flag to pull image before running
As a user who heavily uses generated systemd units in lieu of
kubernetes, but also wants containers to auto-update if I specify a tag
such as `:latest` or a tag that gets re-tagged with updates I was adding
an ExecStartPre option to each unit to pull in these changes.

To optimize this, this new `--pull` flag was added in order to insert
the ExecStartPre option into the unit when used in conjunction with
`--new`.

Signed-off-by: Jacob Lindgren <jlindgre@redhat.com>