increase timeout of on-failure=kill system test
The on-failure=kill system tests turned out to be flaky.
Once the container has been killed, the test waits for
systemd to restart the service by running `container inspect`
for 10 seconds.  The subsequent `healthcheck run` was the
flake point which suggests the 10 seconds timeout to not be
sufficiently high enough; presumably when the CI nodes are
under pressure.

Fixes: #16075
Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>