Remove dangling layers
When pulling an image, layers are unpacked as soon as they (and their
parents) are ready. If we now interrupt this process, a layer may exists
that is not associated with any image, as not all layers for that image
have been pulled yet.

This is not really a problem when we later resume pulling the image.
When resuming we will reuse the existing layer and attach to the image
once done.

However if we don't resume then there is no way to remove this dangling
layer (aside from the nuclear `system reset` way).

This patch adds an extra pass to the image prune operation, to also
remove all layers that are not referenced. This way was chosen as layers
itself are not exposed to the API/user as a resource.

Signed-off-by: Christian Speich <cspeich@emlix.com>