Cirrus: Remove build-push tasks
These tasks only ever run on the main branch.  For whatever reason, the
meta task is throwing errors trying to update a non-existant
`build-push-${IMAGE_SUFFIX}` image.  I have no idea how/why this image
got pruned, it shouldn't have.  However, it appears so far to be a
one-off, and it's not actually needed.  Remove reference to it and
the related useless tasks on this branch.

Signed-off-by: Chris Evich <cevich@redhat.com>