GHA: Fix undefined secret env. var.
Because in github-actions, setting a secret variable isn't enough.  You
ALSO have to set it again in your YAML.  I guess it's assumed in the
name of "security" that the person with access to secrets, might not
also have access to update YAML.  Crazy!

Also, while I'm at it.  Bump up the execution schedule WRT the
check_cirrus_cron workflow - this will give re-run jobs more time to
complete.

Signed-off-by: Chris Evich <cevich@redhat.com>