Make a consistently-named windows installer
In order to maintain a static webpage's link to latest windows
installer build, a consistent file name is required.  In addition to
producing a `podman-vX.Y.Z.msi` file, also produce a `podman.msi` file.
Retain the versioned file in case somebody depends on it's presence in
the artifacts archive.

Signed-off-by: Chris Evich <cevich@redhat.com>