Pick a random port on the host for play kube deployments
podman play kube started to expose the port of containers in k8s deployments on
the same port on the host with
https://github.com/containers/podman/pull/15946. However, that breaks once
replicas are involved, as they would then bind to the same port on the host.

With this commit, we change the default behavior so that podman will pick a
random port for the container instead of using `containerPort`, but only if no
`hostPort` is set.

This fixes: https://github.com/containers/podman/issues/16765

Signed-off-by: Dan Čermák <dcermak@suse.com>