test/system/250-systemd.bats: fix flake
Fix a flake in the kube-template test.  After stopping the service, we
want to make sure that the service container gets removed.  However,
ther is a small race window. `systemctl stop` will return when the
service container _exits_.  In between that and the `container exists`
check, the service container may have not yet been removed.  Hence, add
a loop to account for that race.

Fixes: #16047
Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>